Big restructuring of marrinator
The goal here is to get rid of Makefiles and to make everything work with
xcconfig files, and Xcode build rules. So far so good. There is a package
directory which has all the marrinator library and a MyApplication.cpp file
which uses the RTC class to blink an LED once per second.

Etherclock will no longer build. But I will split that out into a separate
repo and use the package from here.